% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeFunctions.R
\name{addWeightByHaul}
\alias{addWeightByHaul}
\title{Calculate total biomass by haul and add to HH-records}
\usage{
addWeightByHaul(d, to1min = TRUE)
}
\arguments{
\item{d}{DATRASraw object}

\item{to1min}{divide by haul duration in minutes? (defaults to TRUE)}
}
\value{
DATRASraw object
}
\description{
Calculate total biomass by haul and add to HH-records
}
\details{
Fits W = a*L^b and applies to all length data (standardized to one minute).
Thus, this only makes sense for one species.
}
