% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALKfuns.R
\name{AIC.ALKmodel}
\alias{AIC.ALKmodel}
\title{Compute Akaike Information Criterion for an age-length key model.}
\usage{
\method{AIC}{ALKmodel}(object, ..., k = 2)
}
\arguments{
\item{object}{an ALKmodel object.}

\item{...}{unused}

\item{k}{Penalty for number of parameters}
}
\value{
AIC value.
}
\description{
Compute Akaike Information Criterion for an age-length key model.
}
