% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALKfuns.R
\name{anova.ALKmodel}
\alias{anova.ALKmodel}
\title{Compute approximate likelihood ratio test for the equality of two age-length keys.}
\usage{
\method{anova}{ALKmodel}(object, object2, ...)
}
\arguments{
\item{object}{An object of class 'ALKmodel'}

\item{object2}{An object of class 'ALKmodel'}

\item{...}{unused}
}
\value{
a p-value.
}
\description{
Compute approximate likelihood ratio test for the equality of two age-length keys.
}
