% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{datras}
\alias{datras}
\alias{datras-package}
\title{Title datras}
\description{
More about what it does (maybe more than one line)
}
\details{
Version 0.01: 14 December 2013
}
\author{
Einar Hjorleifsson
}
\seealso{
link to other documents
}

