% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALKfuns.R
\name{fitALK}
\alias{fitALK}
\title{Fit a continuation-ratio logit model for age given length and possibly other covariates.}
\usage{
fitALK(
  x,
  minAge,
  maxAge,
  mc.cores = 1,
  model = c("cra~LngtCm", "cra~poly(LngtCm,2)",
    "cra~LngtCm+s(lon,lat,bs='ts')")[method],
  method = 1,
  autoChooseK = FALSE,
  useBIC = FALSE,
  varCof = FALSE,
  maxK = 100,
  gamma = 1.4,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{a DATRASraw object.}

\item{minAge}{minimum age group to consider}

\item{maxAge}{maximum age group to consider}

\item{mc.cores}{The number of cores to use, i.e. how many processes will be spawned (at most)}

\item{model}{Model formula(string) for ALK, or a vector of strings specifying the formula for each logit (i.e. number of age groups minus one).}

\item{method}{Use default formula: 1=Normal distributed age-at-length (1st order), 2=Second order polynomial, 3=Spatial ALK, first order length-effect.}

\item{autoChooseK}{Automatic choice of the max. dimension for the basis used to represent the smooth term for spatial ALK. See ?s in the mgcv-package.}

\item{useBIC}{Use Bayesian Information Criterion for smoothness selection instead of AIC.}

\item{varCof}{Use varying coefficients model for spatial effect.}

\item{maxK}{Maximum k to use. Only applies if autoChooseK is TRUE.}

\item{gamma}{Multiplier for AIC score (see ?gam)}

\item{verbose}{Print details about the fitting process.}

\item{...}{Optional extra arguments to gam()}
}
\value{
An object of class 'ALKmodel'
}
\description{
Fit a continuation-ratio logit model for age given length and possibly other covariates.
}
