% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_hl_data}
\alias{get_hl_data}
\title{get_hl_data}
\usage{
get_hl_data(survey = "NS-IBTS", year = 1996, quarter = 1)
}
\arguments{
\item{survey}{Character, name of the survey}

\item{year}{Integer}

\item{quarter}{Integer, values 1 through 4}
}
\description{
Get DATRAS xx data. The function is a
wrapper around the ICES DATRAS webservice API
\href{https://datras.ices.dk/WebServices/DATRASWebService.asmx?op=getCAdata}{getHLdata}
}

