% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datras.R
\name{minus9toNA}
\alias{minus9toNA}
\title{Replace -9 with NA in numeric columns}
\usage{
minus9toNA(x)
}
\arguments{
\item{x}{DATRASraw object}
}
\value{
DATRASraw object
}
\description{
Replace -9 with NA in numeric columns
}
