% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datras.R
\name{plus999toNA}
\alias{plus999toNA}
\title{Replace 999 with NA in numeric columns}
\usage{
plus999toNA(x)
}
\arguments{
\item{x}{DATRASraw object}
}
\value{
DATRASraw object
}
\description{
Replace 999 with NA in numeric columns
}
