% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datras.R
\name{renameDATRAS}
\alias{renameDATRAS}
\title{Rename DATRAS columns}
\usage{
renameDATRAS(x)
}
\arguments{
\item{x}{part of a DATRASraw object (CA, HH or HL)}
}
\value{
part of a DATRASraw object (CA, HH or HL) with renamed columns
}
\description{
Rename some columns to the old names
}
