% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATRAS.R
\name{DATRAS-internal}
\alias{DATRAS-internal}
\alias{addExtraVariables}
\alias{addHaulID}
\alias{addSpectrum}
\alias{checkSpectrum}
\alias{fixMissingHaulIds}
\alias{getAccuracyCM}
\alias{length.DATRASraw}
\alias{print.DATRASraw}
\alias{readICES}
\alias{refactorHaulLevels}
\alias{reorderTimeLevels}
\alias{summary.DATRASraw}
\alias{tryCatch.W.E}
\alias{xtraVars}
\alias{$.DATRASraw}
\alias{$<-.DATRASraw}
\title{Internal DATRAS Functions}
\description{
Internal DATRAS functions
}
\details{
These are not to be called by the user (or in some cases are just 
waiting for proper documentation to be written :).
}
