\name{downloadExchange}
\alias{downloadExchange}
\title{Interface to DATRAS php download script}
\usage{
  downloadExchange(survey, years = NULL)
}
\arguments{
  \item{survey}{}

  \item{years}{}
}
\value{
  NULL
}
\description{
  Interface to DATRAS php download script
}
\details{
  Note: Command line php must be installed.
}

