% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datras.R
\name{getDatrasExchange}
\alias{getDatrasExchange}
\title{Read exchange data into R.}
\usage{
getDatrasExchange(
  survey,
  years,
  quarters,
  strict = TRUE,
  download.hl = TRUE,
  download.ca = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{survey}{Name of survey to download (For a list of available names run \code{icesDatras::getSurveyList()})}

\item{years}{Vector of years to download.}

\item{quarters}{Vector of quarters to download.}

\item{strict}{if TRUE, missing haul ids in age data should be unqiuely matched when filled in, if FALSE a random match will be assigned.}

\item{download.hl}{if FALSE, length frequency data in DATRAS ("HL") are
not downloaded. Default: TRUE}

\item{download.ca}{if FALSE, biological samples in DATRAS ("CA") are
not downloaded. Default: TRUE}

\item{verbose}{Print information? Default: TRUE}
}
\value{
DATRASraw object.
}
\description{
Download raw exchange data from DATRAS web services.
}
\details{
This function downloads raw exchange data and converts it to a DATRASraw object.
If multiple surveys are supplied, all files are read and combined to
a single object.
}
