\name{plotALKfit}
\alias{plotALKfit}
\title{plot fitted ALK for a given haul}
\usage{
  plotALKfit(x, row, lwd = 2, type = "l",
    ylab = "Proportion", xlab = "Cm", ...)
}
\arguments{
  \item{x}{An object of class 'ALKmodel'}

  \item{row}{row number in hydro data (i.e. haul number)}

  \item{lwd}{argument to matplot()}

  \item{type}{argument to matplot()}

  \item{ylab}{argument to matplot()}

  \item{xlab}{argument to matplot()}

  \item{...}{extra parameters for matplot()}
}
\value{
  nothing
}
\description{
  plot fitted ALK for a given haul
}

