% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datras.R
\name{readExchangeDir}
\alias{readExchangeDir}
\title{Read all exchange files in a folder.}
\usage{
readExchangeDir(path = ".", pattern = ".zip", strict = TRUE)
}
\arguments{
\item{path}{File path.}

\item{pattern}{Pattern that the exchange files match.}

\item{strict}{if TRUE, missing haul ids in age data should be unqiuely matched when filled in, if FALSE a random match will be assigned.}
}
\value{
DATRASraw object
}
\description{
Read all exchange files in a folder and combine to a single DATRASraw
object
}
\details{
Calls \code{readExchange} on all the exchange files in the folder and
combines the results.
}
